Ifpack: Modify Ifpack_IC to take the parameter "fact: ict level-of-fill" which is the fill ratio as in Ifpack_ICT. The default fill ratio is now 1.0 which gives a reasonable default behavior. Previously, the default was to keep no nonzeros but the diagonal, which was crazy and caused lots of trouble. Note the parameter "fact: level-of-fill" is meaningless since Ifpack_IC cannot compute IC(k). Maybe Ifpack_IC should be renamed Ifpack_ICTCrout and a true IC(k) method should take over the Ifpack_IC class.

+++ packages/ifpack/src/Ifpack_IC.h
@@ -332,7 +332,7 @@ class Ifpack_IC: public Ifpack_Preconditioner {
   double Athresh_;
   double Rthresh_;
   double Droptol_;
-  int Lfil_;
+  double Lfil_;
 
   void * Aict_;
   void * Lict_;
